+++ dart/example/event_example.dart
@@ -9,7 +9,7 @@ final event = SentryEvent(
   tags: const <String, String>{'project-id': '7371'},
   extra: const <String, String>{'section': '1'},
   fingerprint: const <String>['example-dart'],
-  user: const User(
+  user: User(
     id: '800',
     username: 'first-user',
     email: 'first@user.lan',